+++ src/addressUtilsByron_test.c
@@ -3,7 +3,7 @@
 #include "addressUtilsByron.h"
 #include "cardano.h"
 #include "hexUtils.h"
-#include "test_utils.h"
+#include "testUtils.h"
 
 #define HD HARDENED_BIP32
