+++ src/bip44_test.c
@@ -2,7 +2,7 @@
 
 #include "cardano.h"
 #include "bip44.h"
-#include "test_utils.h"
+#include "testUtils.h"
 
 #define HD HARDENED_BIP32
