+++ src/endian_test.c
@@ -1,7 +1,7 @@
 #ifdef DEVEL
 
 #include "endian.h"
-#include "test_utils.h"
+#include "testUtils.h"
 
 void run_read_test()
 {
