+++ src/hexUtils.c
@@ -1,5 +1,5 @@
 #include "common.h"
-#include "test_utils.h"
+#include "testUtils.h"
 
 uint8_t hex_parseNibble(const char c)
 {
